django/requirements.txt
Django
djangorestframework
psycopg2
django-cors-headers
pika
drf-spectacular
Pillow
gunicorn
